+++ test/Elastica/Aggregation/ChildrenTest.php
@@ -15,7 +15,7 @@ class ChildrenTest extends BaseAggregationTest
     {
         $client = $this->_getClient();
         $index = $client->getIndex('testaggregationchildren');
-        $index->create(['index' => ['number_of_shards' => 2, 'number_of_replicas' => 1]], true);
+        $index->create(['settings' => ['index' => ['number_of_shards' => 2, 'number_of_replicas' => 1]]], true);
 
         $type = $index->getType(\strtolower(
             'typechildren'.\uniqid()