+++ test/Elastica/ClientTest.php
@@ -353,8 +353,6 @@ class ClientTest extends BaseTest
 
         $index = $this->_createIndex(null, true, 2);
 
-        // Create the index, deleting it first if it already exists
-        $index->create([], true);
         $type = $index->getType('_doc');
 
         // Adds 1 document to the index
@@ -771,7 +769,7 @@ class ClientTest extends BaseTest
         $script->setUpsert(['field1' => 'value1', 'field2' => 10, 'field3' => 'should be removed', 'field4' => 'value4']);
 
         // should use document fields because document does not exist, script is avoided
-        $client->updateDocument(1, $script, $index->getName(), $type->getName(), ['fields' => '_source']);
+        $client->updateDocument(1, $script, $index->getName(), $type->getName());
 
         $document = $type->getDocument(1);
 
@@ -787,7 +785,7 @@ class ClientTest extends BaseTest
         $this->assertEquals('value4', $data['field4']);
 
         // should use script because document exists, document values are ignored
-        $client->updateDocument(1, $script, $index->getName(), $type->getName(), ['fields' => '_source']);
+        $client->updateDocument(1, $script, $index->getName(), $type->getName());
 
         $document = $type->getDocument(1);
 
@@ -845,7 +843,7 @@ class ClientTest extends BaseTest
         $newDocument = new Document(1, ['field1' => 'value1updated', 'field2' => 'value2updated']);
         $upsert = new Document(1, ['field1' => 'value1', 'field2' => 'value2']);
         $newDocument->setUpsert($upsert);
-        $client->updateDocument(1, $newDocument, $index->getName(), $type->getName(), ['fields' => '_source']);
+        $client->updateDocument(1, $newDocument, $index->getName(), $type->getName());
 
         $document = $type->getDocument(1);
         $this->assertInstanceOf(Document::class, $document);
@@ -856,7 +854,7 @@ class ClientTest extends BaseTest
         $this->assertEquals('value2', $data['field2']);
 
         // should use update document because document exists, upsert document values are ignored
-        $client->updateDocument(1, $newDocument, $index->getName(), $type->getName(), ['fields' => '_source']);
+        $client->updateDocument(1, $newDocument, $index->getName(), $type->getName());
 
         $document = $type->getDocument(1);
         $this->assertInstanceOf(Document::class, $document);
@@ -886,7 +884,7 @@ class ClientTest extends BaseTest
 
         $newDocument = new Document(1, ['field1' => 'value1', 'field2' => 'value2']);
         $newDocument->setDocAsUpsert(true);
-        $client->updateDocument(1, $newDocument, $index->getName(), $type->getName(), ['fields' => '_source']);
+        $client->updateDocument(1, $newDocument, $index->getName(), $type->getName());
 
         $document = $type->getDocument(1);
         $this->assertInstanceOf(Document::class, $document);
@@ -1046,8 +1044,7 @@ class ClientTest extends BaseTest
             1,
             $script,
             $index->getName(),
-            $type->getName(),
-            ['fields' => '_source']
+            $type->getName()
         );
 
         $data = $type->getDocument(1)->getData();
@@ -1068,8 +1065,7 @@ class ClientTest extends BaseTest
             1,
             $script,
             $index->getName(),
-            $type->getName(),
-            ['fields' => 'field2, field4']
+            $type->getName()
         );
 
         $document = $type->getDocument(1);
@@ -1196,7 +1192,7 @@ class ClientTest extends BaseTest
         $response = $client->request($path, Request::GET, $query);
         $responseArray = $response->getData();
 
-        $this->assertEquals(1, $responseArray['hits']['total']);
+        $this->assertEquals(1, $responseArray['hits']['total']['value']);
     }
 
     /**
@@ -1219,7 +1215,7 @@ class ClientTest extends BaseTest
         $response = $client->request($path, Request::GET, $query);
         $responseArray = $response->getData();
 
-        $this->assertEquals(1, $responseArray['hits']['total']);
+        $this->assertEquals(1, $responseArray['hits']['total']['value']);
     }
 
     /**
@@ -1466,7 +1462,7 @@ class ClientTest extends BaseTest
         $response = $client->requestEndpoint($endpoint);
         $responseArray = $response->getData();
 
-        $this->assertEquals($totalHits, $responseArray['hits']['total']);
+        $this->assertEquals($totalHits, $responseArray['hits']['total']['value']);
     }
 
     public function endpointQueryRequestDataProvider()