+++ test/Elastica/Query/MultiMatchTest.php
@@ -186,15 +186,17 @@ class MultiMatchTest extends BaseTest
         $index = $client->getIndex('test');
 
         $index->create([
-            'analysis' => [
-                'analyzer' => [
-                    'noStops' => [
-                        'type' => 'standard',
-                        'stopwords' => '_none_',
-                    ],
-                    'stops' => [
-                        'type' => 'standard',
-                        'stopwords' => ['not'],
+            'settings' => [
+                'analysis' => [
+                    'analyzer' => [
+                        'noStops' => [
+                            'type' => 'standard',
+                            'stopwords' => '_none_',
+                        ],
+                        'stops' => [
+                            'type' => 'standard',
+                            'stopwords' => ['not'],
+                        ],
                     ],
                 ],
             ],