+++ test/Elastica/TypeTest.php
@@ -550,7 +550,7 @@ class TypeTest extends BaseTest
     {
         $client = $this->_getClient();
         $index = $client->getIndex('zero');
-        $index->create(['index' => ['number_of_shards' => 1, 'number_of_replicas' => 0]], true);
+        $index->create(['settings' => ['index' => ['number_of_shards' => 1, 'number_of_replicas' => 0]]], true);
 
         $docs = [];
         $docs[] = new Document(1, ['id' => 1, 'email' => 'test@test.com', 'username' => 'farrelley']);
@@ -590,7 +590,7 @@ class TypeTest extends BaseTest
         $type->addDocument(new Document($id, ['name' => 'bruce wayne batman', 'counter' => 1]));
         $newName = 'batman';
 
-        $document = new Document();
+        $document = ['name' => 'bruce wayne batman', 'counter' => 1];
         $script = new Script(
             'ctx._source.name = params.name; ctx._source.counter += params.count',
             [
@@ -620,7 +620,7 @@ class TypeTest extends BaseTest
         $type->addDocument(new Document($id, ['name' => 'bruce wayne batman', 'counter' => 1]));
         $newName = 'batman';
 
-        $document = new Document();
+        $document = ['name' => 'bruce wayne batman', 'counter' => 1];
         $script = new Script(
             'ctx._source.name = params.name; ctx._source.counter += params.count',
             [
@@ -650,7 +650,7 @@ class TypeTest extends BaseTest
         $type->addDocument(new Document($id, ['name' => 'bruce wayne batman', 'counter' => 1]));
         $newName = 'batman';
 
-        $document = new Document();
+        $document = ['name' => 'bruce wayne batman', 'counter' => 1];
         $script = new Script(
             'ctx._source.name = params.name; ctx._source.counter += params.count',
             [
@@ -669,7 +669,7 @@ class TypeTest extends BaseTest
 
             $this->assertContains('action_request_validation_exception', $error['type']);
             $this->assertContains('can\'t provide version in upsert request', $error['reason']);
-            $this->assertContains('can\'t provide both upsert request and a version', $error['reason']);
+            $this->assertContains('Validation Failed: 1: can\'t provide version in upsert request;', $error['reason']);
         }
         $updatedDoc = $type->getDocument($id)->getData();
 
@@ -960,7 +960,7 @@ class TypeTest extends BaseTest
         $mapping = new Mapping($type, $expect = [
             'id' => ['type' => 'integer', 'store' => true],
         ]);
-        $type->setMapping($mapping);
+        $type->setMapping($mapping, ['include_type_name' => true]);
 
         $client = $index->getClient();
 
@@ -984,7 +984,7 @@ class TypeTest extends BaseTest
         $mapping = new Mapping($type, $expect = [
             'id' => ['type' => 'integer', 'store' => true],
         ]);
-        $type->setMapping($mapping);
+        $type->setMapping($mapping, ['include_type_name' => true]);
 
         $client = $index->getClient();
 
@@ -1005,13 +1005,13 @@ class TypeTest extends BaseTest
         $mapping = new Mapping($type, $expect = [
             'id' => ['type' => 'integer', 'store' => true],
         ]);
-        $type->setMapping($mapping);
+        $type->setMapping($mapping, ['include_type_name' => true]);
 
         $endpoint = new Get();
         $endpoint->setIndex('nonExistsIndex');
         $endpoint->setType('nonExistsType');
 
-        $response = $type->requestEndpoint($endpoint);
+        $response = $type->requestEndpoint($endpoint, true);
         $data = $response->getData();
         $mapping = \array_shift($data);